backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
faiss-cpu==1.7.4
sentence-transformers==2.7.0
paddleocr==2.7.0.3
paddlepaddle==3.0.0
openai==1.3.5
pillow==10.1.0
python-dotenv==1.0.0
numpy==1.24.3
